# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_students_import_students": {
        "properties": {
          "dry_run": {
            "default": true,
            "title": "Dry Run",
            "type": "boolean"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "CSV file",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_students_import_students",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportResultOut": {
        "properties": {
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportError_"
            },
            "title": "Errors",
            "type": "array"
          },
          "guardians_created": {
            "title": "Guardians Created",
            "type": "integer"
          },
          "guardians_linked": {
            "title": "Guardians Linked",
            "type": "integer"
          },
          "students_created": {
            "title": "Students Created",
            "type": "integer"
          },
          "total_rows": {
            "title": "Total Rows",
            "type": "integer"
          },
          "valid_rows": {
            "title": "Valid Rows",
            "type": "integer"
          }
        },
        "required": [
          "dry_run",
          "total_rows",
          "valid_rows",
          "errors",
          "students_created",
          "guardians_created",
          "guardians_linked"
        ],
        "title": "ImportResultOut",
        "type": "object"
      },
...
      "LoginSlip": {
        "properties": {
          "admission_no": {
            "title": "Admission No",
            "type": "string"
          },
          "class_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Name"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          },
          "temporary_password": {
            "title": "Temporary Password",
            "type": "string"
          }
        },
        "required": [
          "student_id",
          "admission_no",
          "full_name",
          "class_name",
          "temporary_password"
        ],
        "title": "LoginSlip",
        "type": "object"
      },
      "LoginSlipsIn": {
        "properties": {
          "student_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Student Ids",
            "type": "array"
          }
        },
        "required": [
          "student_ids"
        ],
        "title": "LoginSlipsIn",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "school_admin",
          "section_head",
          "bursar",
          "teacher",
          "counsellor",
          "parent",
          "student"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "StaffCreateIn": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Full Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "section_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          }
        },
        "required": [
          "full_name",
          "email",
          "role"
        ],
        "title": "StaffCreateIn",
        "type": "object"
      },
      "StaffCreatedOut": {
        "properties": {
          "staff": {
            "$ref": "#/components/schemas/StaffOut"
          },
          "temporary_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporary Password"
          }
        },
        "required": [
          "staff",
          "temporary_password"
        ],
        "title": "StaffCreatedOut",
        "type": "object"
      },
...
      "StaffOut": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/StaffRoleOut"
            },
            "title": "Roles",
            "type": "array"
          },
          "two_factor_enabled": {
            "title": "Two Factor Enabled",
            "type": "boolean"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "full_name",
          "email",
          "roles",
          "two_factor_enabled",
          "is_active"
        ],
        "title": "StaffOut",
        "type": "object"
      },
...
      "StudentDetail": {
        "properties": {
          "admission_no": {
            "title": "Admission No",
            "type": "string"
          },
          "arm_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arm Id"
          },
          "arm_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arm Name"
          },
          "class_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "guardians": {
            "items": {
              "$ref": "#/components/schemas/GuardianOut"
            },
            "title": "Guardians",
            "type": "array"
          },
          "has_login": {
            "title": "Has Login",
            "type": "boolean"
          },
          "house": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "House"
          },
          "house_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "House Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "status": {
            "$ref": "#/components/schemas/StudentStatus"
          }
        },
        "required": [
          "id",
          "admission_no",
          "full_name",
          "gender",
          "class_name",
          "arm_name",
          "arm_id",
          "house",
          "has_login",
          "status",
          "first_name",
          "middle_name",
          "last_name",
          "date_of_birth",
          "house_id",
          "guardians"
        ],
        "title": "StudentDetail",
        "type": "object"
      },
      "StudentIn": {
        "properties": {
          "admission_no": {
            "maxLength": 30,
            "minLength": 1,
            "title": "Admission No",
            "type": "string"
          },
          "arm_id": {
            "format": "uuid",
            "title": "Arm Id",
            "type": "string"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "guardian": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GuardianIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "house_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "House Id"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "middle_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          }
        },
        "required": [
          "admission_no",
          "first_name",
          "last_name",
          "arm_id"
        ],
        "title": "StudentIn",
        "type": "object"
      },
...
      "StudentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudentRow"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "StudentPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DigitalLearning360 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/staff": {
      "get": {
        "operationId": "staff_list_staff",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StaffOut"
                  },
                  "title": "Response Staff List Staff",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Staff",
        "tags": [
          "staff"
        ]
      },
      "post": {
        "operationId": "staff_add_staff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffCreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Staff",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/staff/{user_id}/roles/{role}": {
      "delete": {
        "operationId": "staff_remove_role",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Role"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Role",
        "tags": [
          "staff"
        ]
      }
    },
    "/api/students": {
      "get": {
        "operationId": "students_list_students",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "arm_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arm Id"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Students",
        "tags": [
          "students"
        ]
      },
      "post": {
        "operationId": "students_create_student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Student",
        "tags": [
          "students"
        ]
      }
    },
    "/api/students/import": {
      "post": {
        "operationId": "students_import_students",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_students_import_students"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Students",
        "tags": [
          "students"
        ]
      }
    },
...
    "/api/students/logins": {
      "post": {
        "description": "Create or reset student logins. Passwords are returned once, for printing slips.",
        "operationId": "students_issue_logins",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginSlipsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LoginSlip"
                  },
                  "title": "Response Students Issue Logins",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Logins",
        "tags": [
          "students"
        ]
      }
    },
    "/api/students/{student_id}": {
      "get": {
        "operationId": "students_get_student",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Student",
        "tags": [
          "students"
        ]
      },
      "patch": {
        "operationId": "students_update_student",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Student",
        "tags": [
          "students"
        ]
      }
    },
...
  }
}